tests/wpt/web-platform-tests/app-history/navigate-event/navigate-history-back-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    let target_key = i.contentWindow.appHistory.current.key;
    let target_id = i.contentWindow.appHistory.current.id;
    i.contentWindow.appHistory.navigate("?foo");
    i.onload = t.step_func(() => {
      i.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
        assert_equals(e.navigationType, "traverse");
        assert_false(e.cancelable);
        assert_false(e.canTransition);
        assert_false(e.userInitiated);
        assert_false(e.hashChange);
        assert_equals(new URL(e.destination.url).pathname, "/common/blank.html");
        assert_false(e.destination.sameDocument);
        assert_equals(e.destination.key, target_key);
        assert_equals(e.destination.id, target_id);
        assert_equals(e.destination.index, 0);
        assert_equals(e.formData, null);
        assert_equals(e.info, undefined);
      });
      assert_true(i.contentWindow.appHistory.canGoBack);
      i.contentWindow.history.back();
    })
  });
}, "navigate event for history.back() - cross-document");
</script>
